backend/requirements.txt
grpcio==1.63.0
grpcio-tools==1.63.0
protobuf==5.26.1